@@ -37,7 +37,7 @@
     "dexie": "^2.0.4",
     "dotenv": "^8.2.0",
     "electron-default-menu": "^1.0.1",
-    "electron-store": "^5.1.1",
+    "electron-store": "^8.0.1",
     "graphql": "^15.5.0",
     "immutability-helper": "^2.7.0",
     "ipc-promise": "^0.1.4",
@@ -49,11 +49,10 @@
     "memoize-one": "^5.0.0",
     "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
-    "notistack": "^0.9.9",
     "prop-types": "^15.6.1",
     "rasterizehtml": "^1.3.0",
     "react": "^16.8.0",
-    "react-avatar": "^2.5.1",
+    "react-avatar": "^3.10.0",
     "react-bootstrap": "^0.31.5",
     "react-bootstrap-typeahead": "^3.1.4",
     "react-dnd": "^2.6.0",
@@ -93,6 +92,7 @@
     "subscriptions-transport-ws": "^0.8.3",
     "tail": "^1.2.3",
     "trace": "^3.1.0",
+    "uuid": "^8.3.2",
     "winston": "^3.2.1"
   },
   "devDependencies": {
@@ -107,7 +107,7 @@
     "cross-env": "^2.0.1",
     "css-loader": "^1.0.0",
     "devtron": "^1.4.0",
-    "electron": "^9.0.5",
+    "electron": "^11.5.0",
     "electron-builder": "^22.7.0",
     "electron-debug": "^3.1.0",
     "electron-notarize": "^1.0.0",
@@ -120,20 +120,21 @@
     "mocha": "^3.5.3",
     "mockery": "^2.1.0",
     "ncp": "^2.0.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^6.0.1",
     "nyc": "^7.1.0",
+    "playwright": "^1.16.3",
     "react-hot-loader": "^3.1.3",
     "rimraf": "^2.6.2",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^10.2.0",
     "sinon": "^1.17.7",
-    "spectron": "^11.0.0",
+    "spectron": "^13.0.0",
     "stack-chain": "^2.0.0",
     "style-loader": "^0.23.0",
     "tape": "^4.9.0",
     "url-loader": "^1.1.1",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.9"
+    "webpack": "^4.46.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.4.0"
   },
   "peerDependencies": {
     "react": "^16.8.0",
@@ -152,7 +153,7 @@
     "docs": "mkdir -p docs && jsdoc -t ./node_modules/minami -d docs -R README.md -r src/",
     "test": "nyc --reporter=lcov --check-coverage --lines=1 --functions=1 --branches=1 babel-node test/",
     "test:seed-dexie": "node ./test/data/seed-collection-data.js",
-    "test:e2e-single": "npm run build:webpack && env NODE_ENV=test mocha --timeout=120000 test/e2e/computation-run.test.js",
+    "test:e2e-single": "npm run build:webpack && env NODE_ENV=test mocha --inspect-brk --timeout=120000 test/e2e/computation-run.test.js",
     "test:e2e": "npm run build:webpack && env NODE_ENV=test mocha --timeout=120000 test/e2e/*.test.js",
     "watch": "webpack-dev-server --mode=development"
   },
