@@ -449,20 +449,20 @@ const resolvers = {
         return Boom.internal(`Failed to fetch the headless client ${id}`, error);
       }
     },
-    fetchAllDatasetsTags: async () => {
+    fetchAllDatasetsSubjectGroups: async () => {
       const db = database.getDbInstance();
 
       const result = await db.collection('datasets').aggregate([
-        { $unwind: '$tags' },
-        { $group: { _id: null, tags: { $addToSet: '$tags' } } },
-        { $unwind: '$tags' },
-        { $sort: { tags: 1 } },
-        { $group: { _id: null, tags: { $push: '$tags' } } },
+        { $unwind: '$otherInfo.subjectGroups' },
+        { $group: { _id: null, subjectGroups: { $addToSet: '$otherInfo.subjectGroups' } } },
+        { $unwind: '$subjectGroups' },
+        { $sort: { subjectGroups: 1 } },
+        { $group: { _id: null, subjectGroups: { $push: '$subjectGroups' } } },
       ]).toArray();
 
-      return result.length ? result[0].tags : [];
+      return result.length ? result[0].subjectGroups : [];
     },
-    searchDatasets: async (parent, { searchString = '', tags = [] }) => {
+    searchDatasets: async (parent, { searchString = '', subjectGroups = [], modality = '' }) => {
       const db = database.getDbInstance();
 
       const searchObj = {};
@@ -473,8 +473,12 @@ const resolvers = {
         };
       }
 
-      if (tags && tags.length) {
-        searchObj.tags = { $all: tags };
+      if (subjectGroups && subjectGroups.length) {
+        searchObj['otherInfo.subjectGroups'] = { $all: subjectGroups };
+      }
+
+      if (modality) {
+        searchObj['otherInfo.modality'] = modality;
       }
 
       const datasets = await db.collection('datasets').find(searchObj).toArray();
@@ -1449,7 +1453,7 @@ const resolvers = {
       }
     },
     saveDataset: async (parent, args, { credentials }) => {
-      const { id, datasetDescription, participantsDescription } = args.input;
+      const { id, datasetDescription, participantsDescription, otherInfo } = args.input;
 
       const db = database.getDbInstance();
 
@@ -1471,6 +1475,7 @@ const resolvers = {
             $set: {
               datasetDescription,
               participantsDescription,
+              otherInfo,
             }
           },
           { returnDocument: 'after' },
@@ -1481,6 +1486,7 @@ const resolvers = {
         const insertResult = await db.collection('datasets').insertOne({
           datasetDescription,
           participantsDescription,
+          otherInfo,
           owner: {
             id: credentials._id,
             username: credentials.username,
