@@ -165,6 +165,7 @@ const typeDefs = `
     id: ID
     datasetDescription: JSON
     participantsDescription: JSON
+    otherInfo: JSON
     owner: JSON
   }
 
@@ -177,6 +178,7 @@ const typeDefs = `
     id: ID
     datasetDescription: JSON
     participantsDescription: JSON
+    otherInfo: JSON
   }
 
   # This is the general mutation description
@@ -232,8 +234,8 @@ const typeDefs = `
     fetchAllThreads: [Thread]
     fetchUsersOnlineStatus: JSON
     fetchAvailableHeadlessClients: [HeadlessClient]
-    fetchAllDatasetsTags: [String]
-    searchDatasets(searchString: String, tags: [String]): [Dataset]
+    fetchAllDatasetsSubjectGroups: [String]
+    searchDatasets(searchString: String, subjectGroups: [String], modality: String): [Dataset]
     fetchDataset(id: ID!): Dataset
   }
 
