@@ -138,12 +138,13 @@ FilePicker.defaultProps = {
   directory: false,
   tooltip: null,
   selectedFiles: [],
+  extensions: [],
 };
 
 FilePicker.propTypes = {
   multiple: PropTypes.bool,
   directory: PropTypes.bool,
-  extensions: PropTypes.array.isRequired,
+  extensions: PropTypes.array,
   filterName: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
   tooltip: PropTypes.string,
