@@ -5,7 +5,7 @@ import { useQuery, ApolloProvider } from '@apollo/client';
 import { ApolloProvider as ApolloHOCProvider } from '@apollo/react-hoc';
 import { get } from 'lodash';
 import PropTypes from 'prop-types';
-import { ipcRenderer, remote } from 'electron';
+import { ipcRenderer } from 'electron';
 import { List, ListItem } from '@material-ui/core';
 import ArrowUpwardIcon from '@material-ui/icons/ArrowUpward';
 
@@ -183,7 +183,7 @@ Dashboard.propTypes = {
   hideTutorial: PropTypes.bool.isRequired,
 };
 
-const clientConfig = remote.getGlobal('config');
+const { apiServer, subApiServer } = window.config;
 
 function ConnectedDashboard(props) {
   const apolloClient = useRef(null);
@@ -199,7 +199,7 @@ function ConnectedDashboard(props) {
   }, []);
 
   if (!apolloClient.current) {
-    apolloClient.current = getApolloClient(clientConfig);
+    apolloClient.current = getApolloClient({ apiServer, subApiServer });
   }
 
   return (
