@@ -1,13 +1,15 @@
 import React, { useState } from 'react';
-import { useLazyQuery, useMutation } from '@apollo/client';
+import { useLazyQuery, useMutation, useQuery } from '@apollo/client';
 import { Link } from 'react-router';
 import { get } from 'lodash';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
 import Fab from '@material-ui/core/Fab';
+import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
 import TextField from '@material-ui/core/TextField';
 import Typography from '@material-ui/core/Typography';
+import Autocomplete from '@material-ui/lab/Autocomplete';
 import Skeleton from '@material-ui/lab/Skeleton';
 import AddIcon from '@material-ui/icons/Add';
 import SearchIcon from '@material-ui/icons/Search';
@@ -18,16 +20,24 @@ import ListDeleteModal from '../common/list-delete-modal';
 import {
   SEARCH_DATASETS_QUERY,
   DELETE_DATASET_MUTATION,
+  FETCH_ALL_DATASETS_SUBJECT_GROUPS_QUERY,
 } from '../../state/graphql/functions';
+
+const DATASET_MODALITY_OPTIONS = ['sMRI', 'fMRI', 'dMRI', 'pMRI'];
 
 function DataDiscovery() {
   const classes = useStyles();
 
   const [executedFirstSearch, setExecutedFirstSearch] = useState(false);
+
   const [searchString, setSearchString] = useState('');
+  const [subjectGroupsFilter, setSubjectGroupsFilter] = useState([]);
+  const [modalityFilter, setModalityFilter] = useState('');
+
   const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [datasetToDelete, setDatasetToDelete] = useState(null);
 
+  const { data: subjectGroupsData } = useQuery(FETCH_ALL_DATASETS_SUBJECT_GROUPS_QUERY);
   const [searchDatasets, { data, loading }] = useLazyQuery(SEARCH_DATASETS_QUERY, { fetchPolicy: 'cache-and-network' });
   const [deleteRemoteDataset] = useMutation(DELETE_DATASET_MUTATION);
 
@@ -35,10 +45,24 @@ function DataDiscovery() {
     setSearchString(e.target.value);
   }
 
+  function handleSubjectGroupsChange(e, value) {
+    setSubjectGroupsFilter(value);
+  }
+
+  function handleModalityChange(e, value) {
+    setModalityFilter(value);
+  }
+
   function search() {
     setExecutedFirstSearch(true);
 
-    searchDatasets({ variables: { searchString } });
+    searchDatasets({
+      variables: {
+        searchString,
+        subjectGroups: subjectGroupsFilter,
+        modality: modalityFilter,
+      },
+    });
   }
 
   function closeDeleteModal() {
@@ -67,6 +91,7 @@ function DataDiscovery() {
   }
 
   const results = get(data, 'searchDatasets', []);
+  const subjectGroups = get(subjectGroupsData, 'fetchAllDatasetsSubjectGroups', []);
 
   return (
     <div>
@@ -84,13 +109,48 @@ function DataDiscovery() {
         </Fab>
       </div>
       <Box marginY={2}>
-        <TextField
-          id="search"
-          label="Search"
-          fullWidth
-          value={searchString}
-          onChange={handleSearchInputChange}
-        />
+        <Grid container spacing={1}>
+          <Grid item xs={12}>
+            <TextField
+              id="search"
+              label="Search"
+              fullWidth
+              value={searchString}
+              onChange={handleSearchInputChange}
+            />
+          </Grid>
+          <Grid item xs={12} md={6}>
+            <Autocomplete
+              multiple
+              id="subject-groups-filter"
+              options={subjectGroups}
+              value={subjectGroupsFilter}
+              onChange={handleSubjectGroupsChange}
+              renderInput={params => (
+                <TextField
+                  {...params}
+                  variant="standard"
+                  label="Subject Groups"
+                />
+              )}
+            />
+          </Grid>
+          <Grid item xs={12} md={6}>
+            <Autocomplete
+              id="modality-filter"
+              options={DATASET_MODALITY_OPTIONS}
+              value={modalityFilter}
+              onChange={handleModalityChange}
+              renderInput={params => (
+                <TextField
+                  {...params}
+                  variant="standard"
+                  label="Modality"
+                />
+              )}
+            />
+          </Grid>
+        </Grid>
         <Button
           variant="contained"
           color="primary"
